Move parsing boolean to another level of abastraction
